Just organising the modules.
Seperated data into a seperate data folder.
Corresponding changes iin the imports.
