+++ src/data/days.js
@@ -1,0 +1,10 @@
+var days = {
+            0: 'Sun',
+            1: 'Mon',
+            2: 'Tue',
+            3: 'Wed',
+            4: 'Thu',
+            5: 'Fri',
+            6: 'Sat'};
+
+export default days;
