+++ src/data/maxDaysOfMonth.js
@@ -1,0 +1,16 @@
+var maxDaysOfMonth = [
+                    31,
+                    28,
+                    31,
+                    30,
+                    31,
+                    30,
+                    31,
+                    31,
+                    30,
+                    31,
+                    30,
+                    31];
+//This is the array of maximum number of days in each month.
+
+export default maxDaysOfMonth;
