+++ src/data/months.js
@@ -1,0 +1,15 @@
+var months = [
+                'Jan',
+                'Feb', 
+                'March',
+                'April',
+                'May',
+                'June',
+                'July',
+                'August',
+                'Sep',
+                'October',
+                'Nov',
+                'Dec'];
+
+export default months;
